[Base] - Reorganização de componentes (objects e sections)
