+++ components/objects/BlogPostThumbnail/BlogPostThumbnail.module.scss
