+++ components/objects/BlogPostThumbnail/index.js
