+++ components/sections/BlogPreview/BlogPreview.module.scss
@@ -1,6 +1,6 @@
 .blogPreviewComponent {
   background-color: #e2e2e2;
-  background-image: url(../../shared/media/images/backgroundBlog.jpg);
+  background-image: url(../../../shared/media/images/backgroundBlog.jpg);
   background-repeat: no-repeat;
   background-size: cover;
   background-position: center;
