+++ components/sections/Footer/Footer.module.scss
@@ -1,6 +1,6 @@
 .footerComponent {
   padding-top: 60px;
-  background-image: url(../../shared/media/images/backgroundFooter.png);
+  background-image: url(../../../shared/media/images/backgroundFooter.png);
   background-position: center top;
   background-size: cover;
   background-repeat: no-repeat;
