+++ components/sections/Header/Header.module.scss
@@ -10,7 +10,7 @@
     #1562fd,
     #5042f4
   );
-  background-image: url(../../shared/media/images/backgroundHeader.png);
+  background-image: url(../../../shared/media/images/backgroundHeader.png);
   background-size: cover;
   background-position: center bottom;
   background-repeat: no-repeat;
