+++ pages/index.js
@@ -1,5 +1,5 @@
-import Header from "../components/Header";
-import BlogPreview from "../components/BlogPreview";
+import Header from "../components/sections/Header";
+import BlogPreview from "../components/sections/BlogPreview";
 
 export default function Index() {
   return (
