+++ themes/Layout/index.js
@@ -1,5 +1,5 @@
-import Navbar from "../../components/Navbar";
-import Footer from "../../components/Footer";
+import Navbar from "../../components/sections/Navbar";
+import Footer from "../../components/sections/Footer";
 
 export default function Layout({ children }) {
   return (
